BACKEND/requirements.txt
django
djangorestframework
openai
transformers
spacy
nltk
pandas
django-cors-headers
mysqlclient
python-decouple
cryptography
python-dotenv
mysql-connector-python
google-generativeai
